feat: Configure MongoDB and PostgreSQL with JPA/Hibernate profiles
- Added MongoDB configuration in `application.properties` for database connection.
- Configured PostgreSQL connection properties in `application.properties`.
- Set up JPA/Hibernate with PostgreSQL dialect for ORM support.
- Created profiles for different environments (dev, prod) to switch between MongoDB and PostgreSQL easily.

Changes applied:
- `application.properties` updated for MongoDB and PostgreSQL connection.
- Spring profiles added for database configuration.
