+++ .idea/sqldialects.xml
@@ -2,5 +2,7 @@
 <project version="4">
   <component name="SqlDialectMappings">
     <file url="file://$PROJECT_DIR$/db_init.sql" dialect="H2" />
+    <file url="file://$PROJECT_DIR$/spring-boot-course/src/main/resources/data.sql" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/spring-boot-course/src/main/resources/schema.sql" dialect="PostgreSQL" />
   </component>
 </project>
