+++ spring-boot-course/src/main/java/code/with/vanilson/after/events/Event.java
@@ -1,0 +1,38 @@
+package code.with.vanilson.after.events;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "events")
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Event {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @ManyToOne(optional = false)
+    @JoinColumn(name = "organizer_id", referencedColumnName = "id", nullable = false)
+    private Organizer organizer;
+
+    @ManyToOne(optional = false)
+    @JoinColumn(name = "venue_id", referencedColumnName = "id", nullable = false)
+    private Venue venue;
+
+    @Column
+    private LocalDate startDate;
+
+    @Column
+    private LocalDate endDate;
+
+}
