+++ spring-boot-course/src/main/java/code/with/vanilson/after/events/EventRepository.java
@@ -1,0 +1,11 @@
+package code.with.vanilson.after.events;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+@Repository
+public interface EventRepository extends JpaRepository<Event, Integer> {
+
+    List<Event> findByOrganizerId(int organizerId);
+}
