+++ spring-boot-course/src/main/java/code/with/vanilson/after/events/Organizer.java
@@ -1,0 +1,25 @@
+package code.with.vanilson.after.events;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "organizers")
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Organizer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column
+    private String description;
+
+}
