+++ spring-boot-course/src/main/java/code/with/vanilson/after/events/ProductRepository.java
@@ -1,0 +1,11 @@
+package code.with.vanilson.after.events;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+    List<Product> findByEventId(int eventId);
+}
