+++ spring-boot-course/src/main/java/code/with/vanilson/after/events/Venue.java
@@ -1,0 +1,31 @@
+package code.with.vanilson.after.events;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "venues")
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+
+public class Venue {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column
+    private String street;
+
+    @Column
+    private String city;
+
+    @Column
+    private String country;
+}
