+++ spring-boot-course/src/main/java/code/with/vanilson/after/events/VenueRepository.java
@@ -1,0 +1,8 @@
+package code.with.vanilson.after.events;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface VenueRepository extends JpaRepository<Venue, Integer> {
+}
