+++ spring-boot-course/src/main/java/code/with/vanilson/after/registration/Registration.java
@@ -1,0 +1,14 @@
+package code.with.vanilson.after.registration;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document("registrations")
+public record Registration(
+        @Id String id,
+        @NotNull(message = "Product id is required") Integer productId,
+        String ticketCode,
+        @NotBlank(message = "Attendee name is required") String attendeeName) {
+}
