+++ spring-boot-course/src/main/java/code/with/vanilson/after/registration/RegistrationRepository.java
@@ -1,0 +1,13 @@
+package code.with.vanilson.after.registration;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+@Repository
+public interface RegistrationRepository extends MongoRepository<Registration, String> {
+
+    Optional<Registration> findByTicketCode(String ticketCode);
+
+    void deleteByTicketCode(String ticketCode);
+}
