+++ spring-boot-course/src/main/java/code/with/vanilson/before/events/Organizer.java
@@ -1,0 +1,14 @@
+package code.with.vanilson.before.events;
+
+/**
+ * Organizer
+ *
+ * @author vamuhong
+ * @version 1.0
+ * @since 2024-12-02
+ */
+
+public record Organizer(
+        int id,
+        String name,
+        String description){}
