+++ spring-boot-course/src/main/java/code/with/vanilson/before/events/Venue.java
@@ -1,0 +1,16 @@
+package code.with.vanilson.before.events;
+
+/**
+ * Organizer
+ *
+ * @author vamuhong
+ * @version 1.0
+ * @since 2024-12-02
+ */
+public record Venue(
+        int id,
+        String name,
+        String street,
+        String city,
+        String country) {
+}
