--- spring-boot-course/src/main/java/code/with/vanilson/event/Event.java
@@ -1,20 +1,0 @@
-package code.with.vanilson.event;
-
-import java.time.LocalDate;
-
-/**
- * Organizer
- *
- * @author vamuhong
- * @version 1.0
- * @since 2024-12-02
- */
-@SuppressWarnings("unused")
-public record Event(
-        int id,
-        String name,
-        Organizer organizer,
-        Venue venue,
-        LocalDate startDate,
-        LocalDate endDate) {
-}
