--- spring-boot-course/src/main/java/code/with/vanilson/event/Organizer.java
@@ -1,15 +1,0 @@
-package code.with.vanilson.event;
-
-/**
- * Organizer
- *
- * @author vamuhong
- * @version 1.0
- * @since 2024-12-02
- */
-
-public record Organizer(
-        int id,
-        String name,
-        String description) {
-}
