--- spring-boot-course/src/main/java/code/with/vanilson/event/Product.java
@@ -1,14 +1,0 @@
-package code.with.vanilson.event;
-
-import java.math.BigDecimal;
-
-/**
- * Organizer
- *
- * @author vamuhong
- * @version 1.0
- * @since 2024-12-02
- */
-@SuppressWarnings("unused")
-public record Product(int id, int eventId, String name, String description, BigDecimal price) {
-}
