--- spring-boot-course/src/main/java/code/with/vanilson/event/Venue.java
@@ -1,16 +1,0 @@
-package code.with.vanilson.event;
-
-/**
- * Organizer
- *
- * @author vamuhong
- * @version 1.0
- * @since 2024-12-02
- */
-public record Venue(
-        int id,
-        String name,
-        String street,
-        String city,
-        String country) {
-}
