--- spring-boot-course/src/main/java/code/with/vanilson/registration/RegistrationController.java
@@ -1,39 +1,0 @@
-package code.with.vanilson.registration;
-
-import jakarta.validation.Valid;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.NoSuchElementException;
-
-@RestController
-@RequestMapping(path = "/api/registrations")
-public class RegistrationController {
-
-    private final RegistrationRepository registrationRepository;
-
-    public RegistrationController(RegistrationRepository registrationRepository) {
-        this.registrationRepository = registrationRepository;
-    }
-
-    @PostMapping(value = "/create")
-    public Registration create(@RequestBody @Valid Registration registration) {
-        return registrationRepository.create(registration);
-    }
-
-    @GetMapping(path = "/{ticketCode}")
-    public Registration get(@PathVariable("ticketCode") String ticketCode) {
-        return registrationRepository.findByTicketCode(ticketCode)
-                .orElseThrow(
-                        () -> new NoSuchElementException("Registration with ticket code " + ticketCode + " not found"));
-    }
-
-    @PutMapping
-    public Registration update(@RequestBody Registration registration) {
-        return registrationRepository.update(registration);
-    }
-
-    @DeleteMapping(path = "/{ticketCode}")
-    public void delete(@PathVariable("ticketCode") String ticketCode) {
-        registrationRepository.deleteByTicketCode(ticketCode);
-    }
-}
